pyproject.toml
[project]
name = "recsys2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.18",
    "jupyterlab>=4.4.2",
    "loguru>=0.7.3",
    "mlcroissant>=1.0.17",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "polars-lts-cpu>=1.29.0",
    "ray[data,serve,train,tune]>=2.46.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "xgboost>=3.0.1",
]



[project.optional-dependencies]
cpu = [
  "torch>=2.7.0",
  "torchvision>=0.22.0",
]
cu128 = [
  "torch>=2.7.0",
  "torchvision>=0.22.0",
]


[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu128" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true


[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "plotnine>=0.14.5",
    "pytest>=8.3.5",
    "ruff>=0.11.10",
]


#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"


[tool.ruff.lint]
extend-select = ["B"]

